Add the ability to create context without error checking.  More cleanup
