+++ tests/test_config_load.cpp
@@ -8,7 +8,7 @@
 
 #include <catch2/catch_test_macros.hpp>
 
-#include <sgct/readconfig.h>
+#include <sgct/config.h>
 #include <filesystem>
 #include <glm/gtc/quaternion.hpp>
 #include <glm/gtc/type_ptr.hpp>
