+++ tests/test_config_required_parameters.cpp
@@ -9,7 +9,7 @@
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/matchers/catch_matchers_exception.hpp>
 
-#include <sgct/readconfig.h>
+#include <sgct/config.h>
 
 
 TEST_CASE("Parse Required: Version", "[parse]") {
