+++ tests/test_config_required_parameters_schema.cpp
@@ -11,33 +11,26 @@
 
 #include <nlohmann/json.hpp>
 #include <nlohmann/json-schema.hpp>
-#include <sgct/readconfig.h>
+#include <sgct/config.h>
 #include <fstream>
 #include <sstream>
 
 namespace {
-    std::string stringify(const std::string& filename) {
-        std::ifstream myfile;
-        myfile.open(filename);
-        std::stringstream buffer;
-        buffer << myfile.rdbuf();
-        return buffer.str();
-    }
-
-    void attemptValidation(const std::string& cfgString) {
-        const std::string schemaString =
-            stringify(std::string(BASE_PATH) + "/sgct.schema.json");
-        const std::filesystem::path schemaDir = std::string(BASE_PATH);
-        sgct::validateConfigAgainstSchema(cfgString, schemaString, schemaDir);
+    void validate(std::string_view cfgString) {
+        const std::string schema = std::string(BASE_PATH) + "/sgct.schema.json";
+        std::string err = sgct::validateConfigAgainstSchema(cfgString, schema);
+        if (!err.empty()) {
+            throw std::runtime_error(err);
+        }
     }
 } // namespace
 
 TEST_CASE("Parse Required Schema: Version", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {}
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At  of {} - required property 'version' not found in object\n"
@@ -46,13 +39,13 @@ TEST_CASE("Parse Required Schema: Version", "[parse schema]") {
 }
 
 TEST_CASE("Parse Required Schema: Cluster/Master Address", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At  of {\"version\":1} - required property 'masteraddress' not found in "
@@ -62,7 +55,7 @@ TEST_CASE("Parse Required Schema: Cluster/Master Address", "[parse schema]") {
 }
 
 TEST_CASE("Parse Required Schema: Node/Address", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -74,7 +67,7 @@ TEST_CASE("Parse Required Schema: Node/Address", "[parse schema]") {
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0 of {\"port\":1} - required property 'address' not found "
@@ -84,7 +77,7 @@ TEST_CASE("Parse Required Schema: Node/Address", "[parse schema]") {
 }
 
 TEST_CASE("Parse Required Schema: Node/Port", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -96,7 +89,7 @@ TEST_CASE("Parse Required Schema: Node/Port", "[parse schema]") {
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0 of {\"address\":\"localhost\"} - required property 'port' "
@@ -106,7 +99,7 @@ TEST_CASE("Parse Required Schema: Node/Port", "[parse schema]") {
 }
 
 TEST_CASE("Parse Required Schema: Window/Size", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -123,7 +116,7 @@ TEST_CASE("Parse Required Schema: Window/Size", "[parse schema]") {
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
           "At /nodes/0/windows/0 of {} - required property 'size' not found in object\n"
@@ -132,7 +125,7 @@ TEST_CASE("Parse Required Schema: Window/Size", "[parse schema]") {
 }
 
 TEST_CASE("Parse Required Schema: FisheyeProjection/Crop/Left", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -162,7 +155,7 @@ TEST_CASE("Parse Required Schema: FisheyeProjection/Crop/Left", "[parse schema]"
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"crop\":{\"bottom\":1.0,"
@@ -173,7 +166,7 @@ TEST_CASE("Parse Required Schema: FisheyeProjection/Crop/Left", "[parse schema]"
 }
 
 TEST_CASE("Parse Required Schema: FisheyeProjection/Crop/Right", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -203,7 +196,7 @@ TEST_CASE("Parse Required Schema: FisheyeProjection/Crop/Right", "[parse schema]
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"crop\":{\"bottom\":1.0,"
@@ -214,7 +207,7 @@ TEST_CASE("Parse Required Schema: FisheyeProjection/Crop/Right", "[parse schema]
 }
 
 TEST_CASE("Parse Required Schema: FisheyeProjection/Crop/Bottom", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -244,7 +237,7 @@ TEST_CASE("Parse Required Schema: FisheyeProjection/Crop/Bottom", "[parse schema
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"crop\":{\"left\":1.0,"
@@ -255,7 +248,7 @@ TEST_CASE("Parse Required Schema: FisheyeProjection/Crop/Bottom", "[parse schema
 }
 
 TEST_CASE("Parse Required Schema: FisheyeProjection/Crop/Top", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -285,7 +278,7 @@ TEST_CASE("Parse Required Schema: FisheyeProjection/Crop/Top", "[parse schema]")
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"crop\":{\"bottom\":1.0,"
@@ -296,7 +289,7 @@ TEST_CASE("Parse Required Schema: FisheyeProjection/Crop/Top", "[parse schema]")
 }
 
 TEST_CASE("Parse Required Schema: PlanarProjection/FOV/Down", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -326,7 +319,7 @@ TEST_CASE("Parse Required Schema: PlanarProjection/FOV/Down", "[parse schema]") 
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of "
@@ -337,7 +330,7 @@ TEST_CASE("Parse Required Schema: PlanarProjection/FOV/Down", "[parse schema]") 
 }
 
 TEST_CASE("Parse Required Schema: PlanarProjection/FOV/Left", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -367,7 +360,7 @@ TEST_CASE("Parse Required Schema: PlanarProjection/FOV/Left", "[parse schema]") 
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"fov\":{\"down\":1.0,"
@@ -379,7 +372,7 @@ TEST_CASE("Parse Required Schema: PlanarProjection/FOV/Left", "[parse schema]") 
 
 
 TEST_CASE("Parse Required Schema: PlanarProjection/FOV/Right", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -409,7 +402,7 @@ TEST_CASE("Parse Required Schema: PlanarProjection/FOV/Right", "[parse schema]")
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"fov\":{\"down\":1.0,"
@@ -420,7 +413,7 @@ TEST_CASE("Parse Required Schema: PlanarProjection/FOV/Right", "[parse schema]")
 }
 
 TEST_CASE("Parse Required Schema: PlanarProjection/FOV/Up", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -450,7 +443,7 @@ TEST_CASE("Parse Required Schema: PlanarProjection/FOV/Up", "[parse schema]") {
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"fov\":{\"down\":1.0,"
@@ -461,7 +454,7 @@ TEST_CASE("Parse Required Schema: PlanarProjection/FOV/Up", "[parse schema]") {
 }
 
 TEST_CASE("Parse Required Schema: ProjectionPlane/LowerLeft", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -488,7 +481,7 @@ TEST_CASE("Parse Required Schema: ProjectionPlane/LowerLeft", "[parse schema]") 
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"type\":"
@@ -500,7 +493,7 @@ TEST_CASE("Parse Required Schema: ProjectionPlane/LowerLeft", "[parse schema]") 
 }
 
 TEST_CASE("Parse Required Schema: ProjectionPlane/UpperLeft", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -527,7 +520,7 @@ TEST_CASE("Parse Required Schema: ProjectionPlane/UpperLeft", "[parse schema]") 
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"lowerleft\":"
@@ -539,7 +532,7 @@ TEST_CASE("Parse Required Schema: ProjectionPlane/UpperLeft", "[parse schema]") 
 }
 
 TEST_CASE("Parse Required Schema: ProjectionPlane/UpperRight", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -566,7 +559,7 @@ TEST_CASE("Parse Required Schema: ProjectionPlane/UpperRight", "[parse schema]")
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"lowerleft\":"
@@ -578,7 +571,7 @@ TEST_CASE("Parse Required Schema: ProjectionPlane/UpperRight", "[parse schema]")
 }
 
 TEST_CASE("Parse Required Schema: SphericalMirror/Mesh", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -603,7 +596,7 @@ TEST_CASE("Parse Required Schema: SphericalMirror/Mesh", "[parse schema]") {
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"type\":"
@@ -614,7 +607,7 @@ TEST_CASE("Parse Required Schema: SphericalMirror/Mesh", "[parse schema]") {
 }
 
 TEST_CASE("Parse Required Schema: SphericalMirror/Mesh/Bottom", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -644,7 +637,7 @@ TEST_CASE("Parse Required Schema: SphericalMirror/Mesh/Bottom", "[parse schema]"
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"mesh\":{\"left\":\"abc\","
@@ -655,7 +648,7 @@ TEST_CASE("Parse Required Schema: SphericalMirror/Mesh/Bottom", "[parse schema]"
 }
 
 TEST_CASE("Parse Required Schema: SphericalMirror/Mesh/Left", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -685,7 +678,7 @@ TEST_CASE("Parse Required Schema: SphericalMirror/Mesh/Left", "[parse schema]") 
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"mesh\":{\"bottom\":"
@@ -697,7 +690,7 @@ TEST_CASE("Parse Required Schema: SphericalMirror/Mesh/Left", "[parse schema]") 
 }
 
 TEST_CASE("Parse Required Schema: SphericalMirror/Mesh/Right", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -727,7 +720,7 @@ TEST_CASE("Parse Required Schema: SphericalMirror/Mesh/Right", "[parse schema]")
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"mesh\":{\"bottom\":"
@@ -739,7 +732,7 @@ TEST_CASE("Parse Required Schema: SphericalMirror/Mesh/Right", "[parse schema]")
 }
 
 TEST_CASE("Parse Required Schema: SphericalMirror/Mesh/Top", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -769,7 +762,7 @@ TEST_CASE("Parse Required Schema: SphericalMirror/Mesh/Top", "[parse schema]") {
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"mesh\":{\"bottom\":"
@@ -781,7 +774,7 @@ TEST_CASE("Parse Required Schema: SphericalMirror/Mesh/Top", "[parse schema]") {
 }
 
 TEST_CASE("Parse Required Schema: SpoutOutputProjection/MappingSpoutName", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -806,7 +799,7 @@ TEST_CASE("Parse Required Schema: SpoutOutputProjection/MappingSpoutName", "[par
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /nodes/0/windows/0/viewports/0/projection of {\"type\":"
@@ -817,7 +810,7 @@ TEST_CASE("Parse Required Schema: SpoutOutputProjection/MappingSpoutName", "[par
 }
 
 TEST_CASE("Parse Required Schema: User/Tracking/Tracker", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -833,7 +826,7 @@ TEST_CASE("Parse Required Schema: User/Tracking/Tracker", "[parse schema]") {
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /users/0/tracking/0 of {\"device\":\"abc\"} - required property "
@@ -843,7 +836,7 @@ TEST_CASE("Parse Required Schema: User/Tracking/Tracker", "[parse schema]") {
 }
 
 TEST_CASE("Parse Required Schema: User/Tracking/Device", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
@@ -859,7 +852,7 @@ TEST_CASE("Parse Required Schema: User/Tracking/Device", "[parse schema]") {
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /users/0/tracking/0 of {\"tracker\":\"abc\"} - required property "
@@ -869,18 +862,17 @@ TEST_CASE("Parse Required Schema: User/Tracking/Device", "[parse schema]") {
 }
 
 TEST_CASE("Parse Required Schema: Tracker/Name", "[parse schema]") {
-    const std::string Sources = R"(
+    constexpr std::string_view Source = R"(
 {
   "version": 1,
   "masteraddress": "localhost",
   "trackers": [
-    {
-    }
+    {}
   ]
 }
 )";
     CHECK_THROWS_MATCHES(
-        attemptValidation(Sources),
+        validate(Source),
         std::exception,
         Catch::Matchers::Message(
             "At /trackers/0 of {} - required property 'name' not found in object\n"
